converted ints to size_t's

+++ stra.c
@@ -16,7 +16,7 @@ size_t Str_getLength(const char pcSrc[]) {
 
 char *Str_copy(char pcDest[], const char pcSrc[]) {
     size_t length = Str_getLength(pcSrc);
-    int i;
+    size_t i;
     assert(pcSrc != NULL);
     assert(pcDest != NULL);
     for (i = 0; i < length; i++) {
@@ -28,7 +28,7 @@ char *Str_copy(char pcDest[], const char pcSrc[]) {
 char *Str_concat(char pcDest[], const char pcSrc[]) {
     size_t lastIndex = Str_getLength(pcDest);
     size_t totalLength = lastIndex + Str_getLength(pcSrc);
-    int i = 0; 
+    size_t i = 0; 
     assert(pcSrc != NULL);
     assert(pcDest != NULL);
     for (lastIndex; lastIndex < totalLength; lastIndex++) {
@@ -43,7 +43,7 @@ size_t Str_compare(const char Str2[], const char Str1[]) {
     size_t str1Length = Str_getLength(Str1);
     size_t str2Length = Str_getLength(Str2);
     size_t max;
-    int i;
+    size_t i;
     assert(Str2 != NULL);
     assert(Str1 != NULL);
     if (str1Length > str2Length) 
@@ -61,7 +61,7 @@ size_t Str_compare(const char Str2[], const char Str1[]) {
 
 char *Str_search(const char pcHaystack[], const char pcNeedle[]) {
    const char* pc = pcHaystack;
-   int i;
+   size_t i;
    size_t haystackLen = Str_getLength(pcHaystack);
    size_t needleLen = Str_getLength(pcNeedle);
    assert(pcHaystack != NULL);
@@ -71,7 +71,7 @@ char *Str_search(const char pcHaystack[], const char pcNeedle[]) {
    }
    for (i = 0; i < haystackLen; i++) {
         if (pcHaystack[i] == pcNeedle[0]) {
-            int j;
+            size_t j;
             for (j = 0; j < needleLen; j++) {
                 if (pcHaystack[i + j] != pcNeedle[j]) {
                     break;
